requirements.txt
google.generativeai
streamlit
opencv-python
python-docx
Pillow
pypdf
